requirements.txt
flask>=2.0.0
requests>=2.25.0
pystray>=0.19.0
Pillow>=9.0.0
pyinstaller>=5.0.0
pywebview>=4.0.0
packaging>=21.0
